+++ Fish the fishes/Scenes/Stats/StatLine.cs
@@ -6,5 +6,5 @@ public partial class StatLine : HBoxContainer
     [Export]
     public Label ScoreLabel { get; set; }
 
-    public uint Score { set { ScoreLabel.Text = value.ToString(); } }
+    public long Score { set { ScoreLabel.Text = value.ToString(); } }
 }
